+++ fragments/mediaplace/detail_field_body_tinymce.php
@@ -32,7 +32,7 @@ if (!empty($field['translatable'])) {
         <div class="mp3-tiny-canvas-row">
             <span class="mp3-lang-badge"><?= rex_escape($badge) ?></span>
             <div class="mp3-tiny-canvas-preview"><?= rex_escape(DetailPanelFormatter::tinyPreviewText($clVal)) ?></div>
-            <button type="button" class="mp3-tiny-canvas-open" data-canvas-field="<?= rex_escape($key) ?>" data-canvas-clang="<?= rex_escape((string) $clang['id']) ?>" data-canvas-label="<?= rex_escape($label . ' (' . $badge . ')') ?>"><i class="fa-solid fa-pen-to-square"></i> Bearbeiten</button>
+            <button type="button" class="mp3-tiny-canvas-open" data-canvas-field="<?= rex_escape($key) ?>" data-canvas-clang="<?= rex_escape((string) $clang['id']) ?>" data-canvas-label="<?= rex_escape($label . ' (' . $badge . ')') ?>"><i class="fa-solid fa-pen-to-square"></i> <?= rex_escape($this->i18n('mediaplace_edit')) ?></button>
         </div>
         <?php
     }
@@ -44,5 +44,5 @@ $tinyVal = (string) $value;
 <input type="hidden" data-json-field="<?= rex_escape($key) ?>" class="mp3-tiny-canvas-value" value="<?= rex_escape($tinyVal) ?>">
 <div class="mp3-tiny-canvas-row">
     <div class="mp3-tiny-canvas-preview"><?= rex_escape(DetailPanelFormatter::tinyPreviewText($tinyVal)) ?></div>
-    <button type="button" class="mp3-tiny-canvas-open" data-canvas-field="<?= rex_escape($key) ?>" data-canvas-label="<?= rex_escape($label) ?>"><i class="fa-solid fa-pen-to-square"></i> Bearbeiten</button>
+    <button type="button" class="mp3-tiny-canvas-open" data-canvas-field="<?= rex_escape($key) ?>" data-canvas-label="<?= rex_escape($label) ?>"><i class="fa-solid fa-pen-to-square"></i> <?= rex_escape($this->i18n('mediaplace_edit')) ?></button>
 </div>