+++ fragments/mediaplace/detail_preview.php
@@ -22,9 +22,9 @@ $filename = $info['filename'];
 <?php if ($info['is_image']): ?>
     <?php $src = DetailPanelFormatter::mediaThumbUrl($filename, 'rex_media_medium', $info['updatedate'], $info['filesize']); ?>
     <div class="mp3-detail-preview">
-        <button type="button" class="mp3-lightbox-open-btn" data-lightbox-src="<?= rex_escape($src) ?>" data-lightbox-caption="<?= rex_escape($info['title'] !== '' ? $info['title'] : $filename) ?>" title="Lightbox öffnen"><i class="fa-solid fa-up-right-and-down-left-from-center"></i></button>
+        <button type="button" class="mp3-lightbox-open-btn" data-lightbox-src="<?= rex_escape($src) ?>" data-lightbox-caption="<?= rex_escape($info['title'] !== '' ? $info['title'] : $filename) ?>" title="<?= rex_escape($this->i18n('mediaplace_open_lightbox')) ?>"><i class="fa-solid fa-up-right-and-down-left-from-center"></i></button>
         <?php if (!empty($info['focuspoint_available'])): ?>
-            <button type="button" class="mp3-focuspoint-edit-btn" data-focuspoint-file="<?= rex_escape($filename) ?>" title="Fokuspunkt bearbeiten"><i class="fa-solid fa-crosshairs"></i></button>
+            <button type="button" class="mp3-focuspoint-edit-btn" data-focuspoint-file="<?= rex_escape($filename) ?>" title="<?= rex_escape($this->i18n('mediaplace_edit_focuspoint')) ?>"><i class="fa-solid fa-crosshairs"></i></button>
         <?php endif; ?>
         <img src="<?= rex_escape($src) ?>" alt="<?= rex_escape($info['title'] !== '' ? $info['title'] : $filename) ?>">
     </div>
@@ -46,7 +46,7 @@ $filename = $info['filename'];
 <?php elseif (DetailPanelFormatter::isPdfFilename($filename)): ?>
     <?php $pdfSrc = DetailPanelFormatter::mediaFileUrl($filename, $info['updatedate'], $info['filesize']); ?>
     <div class="mp3-detail-preview mp3-detail-preview-icon">
-        <a class="mp3-pdf-open-btn" href="<?= rex_escape($pdfSrc) ?>" target="_blank" rel="noopener" title="PDF in neuem Tab öffnen"><i class="fa-solid fa-up-right-from-square"></i></a>
+        <a class="mp3-pdf-open-btn" href="<?= rex_escape($pdfSrc) ?>" target="_blank" rel="noopener" title="<?= rex_escape($this->i18n('mediaplace_open_pdf_new_tab')) ?>"><i class="fa-solid fa-up-right-from-square"></i></a>
         <i class="<?= rex_escape(DetailPanelFormatter::fileIconClass($filename)) ?>"></i>
     </div>
 <?php else: ?>